(121) Relocating archive methods
